+++ RELEASES.md
@@ -1,7 +1,7 @@
 # Next
 
-- Removes `Enum` trait, in flavor of implementing `FromSql`/`ToSql` traits
-    directly;
+- Removes `Composite` and `Enum` traits, in flavor of implementing
+    `FromSql`/`ToSql` traits directly;
 - Adds arbitrary crate support;
 - Updates time to 0.3;
 - Makes pq::Result thread safe;